Teste cronometro

+++ app/src/main/java/com/example/corra/Model/Corrida.java
@@ -10,7 +10,7 @@ public class Corrida {
     public int uid;
 
     @ColumnInfo(name = "tempo")
-    public float tempo;
+    public long tempo;
 
     @ColumnInfo(name = "velocidade")
     public double velocidade;
